generate_request_id fixes (#27)
* Fix generate_request_id with thread_safe

The previous implementation dropped the first guard
before taking another, which could race.

It would be better to change to NonZeroU32 and checked_add as well
(wrapping issues).

* Change generate_request_id to return NonZeroU32

Zero is for broadcast request_ids, change types to make it impossible
to return zero.  i32 wrapping does take a lot of requests, but this is neater.